% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version2.R, R/Version3.R
\name{multiwcomp.MiRKAT}
\alias{multiwcomp.MiRKAT}
\alias{multiwcomp.MiRKAT}
\title{Taxa importance evaluation through weighted MiRKAT}
\usage{
multiwcomp.MiRKAT(x, y, w.seq = c(0.1, 0.4, 0.7, 1, 2, 3, 4, 5), cov = NULL,
  zero.rep = T, pseudocount = 1)

multiwcomp.MiRKAT(x, y, w.seq = c(0.1, 0.4, 0.7, 1, 2, 3, 4, 5), cov = NULL,
  zero.rep = T, pseudocount = 1)
}
\arguments{
\item{x}{matrix with the composition for each individual}

\item{y}{response variable: dichotomous or continuous}

\item{w.seq}{vector of weights to consider in the analysis}

\item{cov}{covariates to hace into account as a possible confounding factors}

\item{zero.rep}{logical value indicating if cmultRepl() frunction from
zCompositions should be used in order to replace null values.}

\item{pseudocount}{a value to add to the whole matrix in order to avoid zeros.
Only if zero.rep == FALSE.}

\item{x}{matrix with the composition for each individual}

\item{y}{response variable: dichotomous or continuous}

\item{w.seq}{vector of weights to consider in the analysis}

\item{cov}{covariates to hace into account as a possible confounding factors}

\item{zero.rep}{logical value indicating if cmultRepl() frunction from
zCompositions should be used in order to replace null values.}

\item{pseudocount}{a value to add to the whole matrix in order to avoid zeros.
Only if zero.rep == FALSE.}
}
\value{
a data.frame with the slope and its p-value for the linear regression
model of wcomp.MiRKAT p-values considering the given weights.

a data.frame with the slope and its p-value for the linear regression
model of wcomp.MiRKAT p-values considering the given weights.
}
\description{
For different weight values, computes for each variable the p-value using
`wcomp.MiRKAT` function and implements a linear regression model giving back
its slope and the p-value for the slope.

For different weight values, computes for each variable the p-value using
`wcomp.MiRKAT` function and implements a linear regression model giving back
its slope and the p-value for the slope.
}
\examples{

# Matrix of samples
 x <- matrix(runif(100), nrow = 10)
# Response vector
 y <- rep(c(0,1),5)

# Run multiwcomp.MiRKAT function
 multiwcomp.MiRKAT(x,y)

# Matrix of samples
 x <- matrix(runif(100), nrow = 10)
# Response vector
 y <- rep(c(0,1),5)

# Run multiwcomp.MiRKAT function
 multiwcomp.MiRKAT(x,y)
}
