% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version2.R, R/Version3.R
\name{w.dist}
\alias{w.dist}
\alias{w.dist}
\title{Computes the weighted distance between two compositions}
\usage{
w.dist(x1, x2)

w.dist(x1, x2)
}
\arguments{
\item{x1}{first composition}

\item{x2}{second composition}

\item{w}{vector of weights}

\item{w}{vector of weights}

\item{x1}{first composition}

\item{x2}{second composition}
}
\value{
a value with the weighted distance between x1 and x2

a value with the weighted distance between x1 and x2
}
\description{
Computes the weighted distance between two compositions

Computes the weighted distance between two compositions
}
\examples{

# Build compositional vectors
  x1 <- prop.table(runif(10))
  x2 <- prop.table(runif(10))

# Vector of weights
  w <- c(rep(1,4),3,3,rep(1,4))

# The distance
   w.dist(w,x1,x2)

# Build compositional vectors
  x1 <- prop.table(runif(10))
  x2 <- prop.table(runif(10))

# Vector of weights
  w <- c(rep(1,4),3,3,rep(1,4))

# The distance
   w.dist(w,x1,x2)
}
