% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version2.R, R/Version3.R
\name{w.geoM}
\alias{w.geoM}
\alias{w.geoM}
\title{Weighted geometric mean}
\usage{
w.geoM(w, x)

w.geoM(w, x)
}
\arguments{
\item{w}{vector of weights of the same length as x.}

\item{x}{vector for which the weighted geometric mean is computed.}

\item{w}{vector of weights of the same length as x.}

\item{x}{vector for which the weighted geometric mean is computed.}
}
\value{
the weighted geometric mean value.

the weighted geometric mean value.
}
\description{
Weighted geometric mean

\code{w.geoM} computes the weighted geometric mean as defined in Egozcue and
Pawlowsky (2016)

Weighted geometric mean

\code{w.geoM} computes the weighted geometric mean as defined in Egozcue and
Pawlowsky (2016)
}
\examples{

# Build compositional vectors
  x <- prop.table(runif(10))
# Vector of weights
  w <- c(2,rep(1,9))

# Weighted geometric mean
  w.geoM(w,x)

# Build compositional vectors
  x <- prop.table(runif(10))
# Vector of weights
  w <- c(2,rep(1,9))

# Weighted geometric mean
  w.geoM(w,x)
}
