% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version2.R, R/Version3.R
\name{wcomp.MiRKAT}
\alias{wcomp.MiRKAT}
\alias{wcomp.MiRKAT}
\title{Weighted MiRKAT}
\usage{
wcomp.MiRKAT(x, y, w = rep(1, nrow(x)), cov = NULL)

wcomp.MiRKAT(x, y, w = rep(1, nrow(x)), cov = NULL)
}
\arguments{
\item{x}{matrix with the composition for each individual}

\item{y}{response variable: dichotomous or continuous}

\item{w}{vector of weights for each taxa}

\item{cov}{covariates to hace into account as a possible confounding factors}

\item{x}{matrix with the composition for each individual}

\item{y}{response variable: dichotomous or continuous}

\item{w}{vector of weights for each taxa}

\item{cov}{covariates to hace into account as a possible confounding factors}
}
\value{
the p-value associated to the compositional data with the corresponding
vector of weights.

the p-value associated to the compositional data with the corresponding
vector of weights.
}
\description{
Runs MiRKAT function using the specified weights for each taxon.

Runs MiRKAT function using the specified weights for each taxon.
}
\examples{

# Matrix of samples
 x <- matrix(runif(100), nrow = 10)
# Response vector
 y <- rep(c(0,1),5)
# Vector of weights
 w <- c(2,rep(1,9))

# Run wcomp.MiRKAT function
 wcomp.MiRKAT(x,y,w)

# Matrix of samples
 x <- matrix(runif(100), nrow = 10)
# Response vector
 y <- rep(c(0,1),5)
# Vector of weights
 w <- c(2,rep(1,9))

# Run wcomp.MiRKAT function
 wcomp.MiRKAT(x,y,w)
}
