% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version2.R, R/Version3.R
\name{wM.barplot}
\alias{wM.barplot}
\alias{wM.barplot}
\title{Plot the results in wcomp.MiRKAT}
\usage{
wM.barplot(tab, col = c("red", "blue"))

wM.barplot(tab, col = c("red", "blue"))
}
\arguments{
\item{tab}{the output of wcomp.MiRKAT function}

\item{col}{two colors for the positive and negatives slopes}

\item{tab}{the output of wcomp.MiRKAT function}

\item{col}{two colors for the positive and negatives slopes}
}
\value{
a barplot representing the slopes for each variable

a barplot representing the slopes for each variable
}
\description{
Graphical representation of the slopes presented in the output
of wcomp.MiRKAT function.

Graphical representation of the slopes presented in the output
of wcomp.MiRKAT function.
}
\examples{

# Matrix of samples
 x <- matrix(runif(100), nrow = 10)
# Response vector
 y <- rep(c(0,1),5)

# Run multiwcomp.MiRKAT function
 A <- multiwcomp.MiRKAT(x,y)
# Represent the results
 wM.barplot(A)

# Matrix of samples
 x <- matrix(runif(100), nrow = 10)
# Response vector
 y <- rep(c(0,1),5)

# Run multiwcomp.MiRKAT function
 A <- multiwcomp.MiRKAT(x,y)
# Represent the results
 wM.barplot(A)
}
